requirements-train.txt
pandas==2.3.2
numpy==2.3.3
torch==2.8.0
torchvision==0.23.0
matplotlib==3.10.6
seaborn==0.13.2
kagglehub==0.3.13
pytest==8.3.3
pytest-cov==4.1.0
httpx==0.27.0
mlflow==3.8.0
pyyaml==6.0.2
